Long Beach panic commit: shooter disabled by requst, gearbox limit switch code not currently in use, as it was causing issues, temp auton modes for various durations, will be removed later.

+++ src/Subsystems/Arm.hpp
@@ -32,7 +32,7 @@ public:
     void SetManualOverride(bool manual);
     bool GetManualOverride() const;
 
-    int32_t GetArmHeightValue() const;
+    double GetArmHeightValue() const;
     double GetArmSpeed() const;
 
     void SetManualCarriagePosition(int direction);
